requirements.txt
numpy>=1.14.0
pydot_ng>=1.0.0
scipy>=1.0.0
click>=6.7
attrdict==2.0.0
pandas==0.21.0
neptune_cli==2.7.1
catboost==0.6.1.1
ipython==6.2.1
PyYAML==3.12
Keras
scikit-learn==0.19.1
gensim==3.3.0
translation==1.0.5
langdetect==1.0.7
nltk==3.2.5
xgboost
google_compute_engine
